moved parse route into a new directory, lib, in the client directory.
